+++ packages/elements-core/src/components/TryIt/index.ts
@@ -1,3 +1,3 @@
-export * from './state';
+export { chosenServerAtom } from './chosenServer';
 export * from './TryIt';
 export * from './TryItWithRequestSamples';
